+++ dll/win32/shell32/precomp.h
@@ -43,7 +43,7 @@
 #include <shlobj_undoc.h>
 
 #include <shlwapi_undoc.h>
-#include <IShellFolder_helpers.h>
+#include <ishellfolder_helpers.h>
 
 #include <shellapi.h>
 #undef ShellExecute
