+++ sdk/include/reactos/ishellfolder_helpers.h
